+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 ## [Unreleased]
 ### Added
 - Basic model.
+- Sync of data with hydrus network.